#39 [feat, rename] Test 환경 Query dsl 설정
